requirements.txt
fastapi
uvicorn
streamlit
requests
pydantic
python-dotenv
Pillow
reportlab
httpx